# contracts/oar-openapi.yaml
openapi: 3.1.0
info:
  title: Organization Autorunner HTTP API
  version: 0.3.0
  description: |
    Canonical workspace HTTP contract for Organization Autorunner.

    Topics and cards are the primary durable resources. Threads remain backing
    infrastructure surfaced through read-only inspection and workspace views.
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /topics/{topic_id}:
    get:
      operationId: getTopic
      summary: Get topic
      tags: [topics]
      x-oar-surface: canonical
      x-oar-command-id: topics.get
      x-oar-cli-path: topics get
      x-oar-why: Resolve one topic and its canonical durable fields.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ topic }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [topics]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/TopicID'
      responses:
        '200':
          description: Topic payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: patchTopic
      summary: Patch topic
      tags: [topics]
      x-oar-surface: canonical
      x-oar-command-id: topics.patch
      x-oar-cli-path: topics patch
      x-oar-why: Update topic state with provenance and optimistic concurrency.
      x-oar-input-mode: json-body
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ topic }`."
      x-oar-error-codes: [auth_required, invalid_request, invalid_token, not_found, conflict]
      x-oar-concepts: [topics, write, concurrency]
      x-oar-stability: beta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTopicRequest'
      parameters:
        - $ref: '#/components/parameters/TopicID'
      responses:
        '200':
          description: Topic patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
  /threads/{thread_id}:
    get:
      operationId: getThread
      summary: Inspect backing thread
      tags: [threads]
      x-oar-surface: canonical
      x-oar-command-id: threads.inspect
      x-oar-cli-path: threads inspect
      x-oar-why: Resolve one backing thread for low-level inspection and diagnostics.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ thread }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [threads, inspection]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/ThreadID'
      responses:
        '200':
          description: Thread payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /docs/{document_id}:
    get:
      operationId: getDocument
      summary: Get document
      tags: [docs]
      x-oar-surface: canonical
      x-oar-command-id: docs.get
      x-oar-cli-path: docs get
      x-oar-why: Resolve a document lineage and its current head revision.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document, revision }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [docs]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: Document payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithRevisionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /boards/{board_id}:
    get:
      operationId: getBoard
      summary: Get board
      tags: [boards]
      x-oar-surface: canonical
      x-oar-command-id: boards.get
      x-oar-cli-path: boards get
      x-oar-why: Resolve canonical board state and summary.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ board, summary }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [boards]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/BoardID'
      responses:
        '200':
          description: Board payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: patchBoard
      summary: Patch board
      tags: [boards]
      x-oar-surface: canonical
      x-oar-command-id: boards.patch
      x-oar-cli-path: boards patch
      x-oar-why: Update board metadata with optimistic concurrency.
      x-oar-input-mode: json-body
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ board }`."
      x-oar-error-codes: [auth_required, invalid_request, invalid_token, not_found, conflict]
      x-oar-concepts: [boards, write, concurrency]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/BoardID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchBoardRequest'
      responses:
        '200':
          description: Board patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
  /boards/{board_id}/cards/{card_id}:
    get:
      operationId: getBoardCard
      summary: Get board-scoped card
      tags: [boards, cards]
      x-oar-surface: canonical
      x-oar-command-id: boards.cards.get
      x-oar-cli-path: boards cards get
      x-oar-why: Resolve a card through its board membership context.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ card }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [boards, cards]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/BoardID'
        - $ref: '#/components/parameters/CardID'
      responses:
        '200':
          description: Card payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /artifacts/{artifact_id}:
    get:
      operationId: getArtifact
      summary: Get artifact metadata
      tags: [artifacts]
      x-oar-surface: canonical
      x-oar-command-id: artifacts.get
      x-oar-cli-path: artifacts get
      x-oar-why: Resolve immutable artifact metadata referenced from timelines and packets.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ artifact }`."
      x-oar-error-codes: [auth_required, invalid_token, not_found]
      x-oar-concepts: [artifacts]
      x-oar-stability: beta
      parameters:
        - $ref: '#/components/parameters/ArtifactID'
      responses:
        '200':
          description: Artifact payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    ErrorResponse:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    TopicID:
      in: path
      name: topic_id
      required: true
      schema: { type: string }
    ThreadID:
      in: path
      name: thread_id
      required: true
      schema: { type: string }
    DocumentID:
      in: path
      name: document_id
      required: true
      schema: { type: string }
...
    BoardID:
      in: path
      name: board_id
      required: true
      schema: { type: string }
    CardID:
      in: path
      name: card_id
      required: true
      schema: { type: string }
    ArtifactID:
      in: path
      name: artifact_id
      required: true
      schema: { type: string }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
      required: [code, message]
...
    TypedRef:
      type: string
      description: Typed ref string such as `topic:roadmap` or `artifact:art_123`.
    Provenance:
      type: object
      properties:
        sources:
          type: array
          items: { type: string }
        notes: { type: string }
        by_field:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
      required: [sources]
...
    Artifact:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        content_type: { type: string }
        content_hash: { type: string }
        refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        summary: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
      required: [id, kind, content_type, content_hash, refs, created_at, created_by]
    Thread:
      type: object
      properties:
        id: { type: string }
        topic_ref: { $ref: '#/components/schemas/TypedRef' }
        title: { type: string }
        status:
          type: string
          enum: [active, archived]
        priority:
          type: string
          enum: [p0, p1, p2, p3]
        tags:
          type: array
          items: { type: string }
        latest_event_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required: [id, title, status, priority, tags, created_at, created_by, updated_at, updated_by, provenance]
    Topic:
      type: object
      properties:
        id: { type: string }
        type:
          type: string
          enum: [initiative, objective, decision, incident, risk, request, note, other]
        status:
          type: string
          enum: [proposed, active, blocked, resolved, archived]
        title: { type: string }
        summary: { type: string }
        owner_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        primary_thread_ref: { $ref: '#/components/schemas/TypedRef' }
        document_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        board_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        related_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required:
        [id, type, status, title, summary, owner_refs, document_refs, board_refs, related_refs,
         created_at, created_by, updated_at, updated_by, provenance]
    Document:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        subject_ref: { $ref: '#/components/schemas/TypedRef' }
        head_revision_ref: { $ref: '#/components/schemas/TypedRef' }
        state: { type: string }
        refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required:
        [id, title, subject_ref, head_revision_ref, state, refs, created_at, created_by, updated_at, updated_by, provenance]
    DocumentRevision:
      type: object
      properties:
        revision_id: { type: string }
        document_ref: { $ref: '#/components/schemas/TypedRef' }
        prev_revision_ref: { $ref: '#/components/schemas/TypedRef' }
        artifact_ref: { $ref: '#/components/schemas/TypedRef' }
        revision_number: { type: integer }
        summary: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
      required: [revision_id, document_ref, artifact_ref, revision_number, created_at, created_by]
    Board:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        status:
          type: string
          enum: [active, paused, archived]
        primary_topic_ref: { $ref: '#/components/schemas/TypedRef' }
        primary_thread_ref: { $ref: '#/components/schemas/TypedRef' }
        document_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        card_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        pinned_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required:
        [id, title, status, document_refs, card_refs, pinned_refs, created_at, created_by, updated_at, updated_by, provenance]
    Card:
      type: object
      properties:
        id: { type: string }
        board_ref: { $ref: '#/components/schemas/TypedRef' }
        topic_ref: { $ref: '#/components/schemas/TypedRef' }
        thread_ref: { $ref: '#/components/schemas/TypedRef' }
        document_ref: { $ref: '#/components/schemas/TypedRef' }
        title: { type: string }
        summary: { type: string }
        column_key:
          type: string
          enum: [backlog, ready, in_progress, blocked, review, done]
        rank: { type: string }
        assignee_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        risk:
          type: string
          enum: [low, medium, high, critical]
        resolution:
          type: string
          enum: [unresolved, completed, canceled, superseded]
        resolution_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        related_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required:
        [id, board_ref, title, summary, column_key, rank, assignee_refs, risk, resolution, resolution_refs,
         related_refs, created_at, created_by, updated_at, updated_by, provenance]
    BoardSummary:
      type: object
      properties:
        card_count: { type: integer }
        cards_by_column:
          type: object
          additionalProperties: { type: integer }
        unresolved_card_count: { type: integer }
        resolved_card_count: { type: integer }
        document_count: { type: integer }
        latest_activity_at: { type: string, format: date-time }
        has_document_refs: { type: boolean }
        document_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
      required: [card_count, cards_by_column, unresolved_card_count, resolved_card_count, document_count]
...
    TopicPatchInput:
      type: object
      properties:
        type:
          type: string
          enum: [initiative, objective, decision, incident, risk, request, note, other]
        status:
          type: string
          enum: [proposed, active, blocked, resolved, archived]
        title: { type: string }
        summary: { type: string }
        owner_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        primary_thread_ref: { $ref: '#/components/schemas/TypedRef' }
        document_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        board_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        related_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        provenance: { $ref: '#/components/schemas/Provenance' }
...
    BoardPatchInput:
      type: object
      properties:
        title: { type: string }
        status:
          type: string
          enum: [active, paused, archived]
        primary_topic_ref: { $ref: '#/components/schemas/TypedRef' }
        primary_thread_ref: { $ref: '#/components/schemas/TypedRef' }
        document_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        pinned_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        provenance: { $ref: '#/components/schemas/Provenance' }
...
    PatchTopicRequest:
      type: object
      properties:
        patch: { $ref: '#/components/schemas/TopicPatchInput' }
        if_updated_at:
          type: string
          format: date-time
      required: [patch]
...
    PatchBoardRequest:
      type: object
      properties:
        patch: { $ref: '#/components/schemas/BoardPatchInput' }
        if_updated_at:
          type: string
          format: date-time
      required: [patch]
...
    TopicResponse:
      type: object
      properties:
        topic: { $ref: '#/components/schemas/Topic' }
      required: [topic]
...
    ThreadResponse:
      type: object
      properties:
        thread: { $ref: '#/components/schemas/Thread' }
      required: [thread]
...
    DocumentWithRevisionResponse:
      type: object
      properties:
        document: { $ref: '#/components/schemas/Document' }
        revision: { $ref: '#/components/schemas/DocumentRevision' }
      required: [document, revision]
...
    BoardResponse:
      type: object
      properties:
        board: { $ref: '#/components/schemas/Board' }
        summary: { $ref: '#/components/schemas/BoardSummary' }
      required: [board, summary]
...
    CardResponse:
      type: object
      properties:
        card: { $ref: '#/components/schemas/Card' }
      required: [card]
...
    ArtifactResponse:
      type: object
      properties:
        artifact: { $ref: '#/components/schemas/Artifact' }
      required: [artifact]
...